+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog is following the recommended format by [keepachangelog](https://k
 ### Added
 
 - Add MCP tools for forms CRUD (`listForms`, `getForm`, `createForm`, `updateForm`, `deleteForm`)
+- Add MCP tools for Identity Profiles (`createIdentityProfile`, `listIdentityProfiles`, `setIdentityProfileMapping`)
 
 ### Changed